Refactor calendar types
