+++ csm_web/frontend/src/components/enrollment_automation/calendar/CalendarDay.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { formatTime } from "../utils";
-import { CalendarEventSingleTime } from "./Calendar";
+import { CalendarEventSingleTime } from "./CalendarTypes";
 
 interface NumberTime {
   day: string;
